Support more configuration and dynamic upscaling
Add config options for the maximum amount of workers, the delays between
scaling up or down, and the amount of seconds for which memory and cpu
stats are kept for processes for more control over this module.

Also adds a calculation so it can scale up faster to respond to a sudden
spike in traffic, based on the target CPU utilization, the current
number of workers, and the available memory.